backend/requirements.txt
fastapi
uvicorn[standard]
sqlalchemy[asyncio]
asyncpg
alembic
python-dotenv
python-multipart
pydantic
pydantic-settings
websockets

# Testing
pytest
pytest-asyncio
httpx
aiosqlite

